Possible changed. Only comment for now. config stuff
I commented out lines I feel should be deleted. I added a comment to the
ends of lines that maybe should be deleted but I am not sure.

If it's not being used and I felt it shouldn't it was fully commented.
Not being used but I feel might be somewhere I don't see have the
comments at the end. I removed the else block so the initialization
checks there is a config file and parses the needed info and sends this
info to the do_spdx class